backend/requirements.txt
argcomplete==2.1.2
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
durationpy==0.5
Flask==2.2.3
hjson==3.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
Pillow==9.4.0
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
urllib3==1.26.15
Werkzeug==2.2.3
zappa==0.56.1
zipp==3.15.0
